+++ test.mjs
@@ -2,6 +2,7 @@
 
 import TestDirector from "test-director";
 
+import testAssertKoaContextRequestGraphQL from "./assertKoaContextRequestGraphQL.test.mjs";
 import testCheckGraphQLSchema from "./checkGraphQLSchema.test.mjs";
 import testCheckGraphQLValidationRules from "./checkGraphQLValidationRules.test.mjs";
 import testCheckOptions from "./checkOptions.test.mjs";
@@ -11,6 +12,7 @@ import testGraphQLAggregateError from "./GraphQLAggregateError.test.mjs";
 
 const tests = new TestDirector();
 
+testAssertKoaContextRequestGraphQL(tests);
 testCheckGraphQLSchema(tests);
 testCheckGraphQLValidationRules(tests);
 testCheckOptions(tests);